requirements.txt
beautifulsoup4
ipython
Mastodon.py
openai